Refactor

+++ 2023-12-05_synth/src/output.rs
@@ -11,7 +11,7 @@ pub fn start() -> anyhow::Result<Box<dyn tinyaudio::BaseAudioOutputDevice>> {
     let volume = 0.2;
 
     let mut oscillator =
-        oscillator(sawtooth, frequency, params.sample_rate as f32);
+        oscillator(sawtooth, frequency, params.sample_rate as u32);
 
     let device = tinyaudio::run_output_device(params, move |data| {
         for samples in data.chunks_mut(params.channels_count) {
@@ -53,14 +53,14 @@ fn sawtooth(t: f32) -> f32 {
 fn oscillator(
     wave: fn(f32) -> f32,
     frequency: f32,
-    sample_rate: f32,
+    sample_rate: u32,
 ) -> Signal {
     let mut t = 0.;
 
     iter::from_fn(move || {
         let value = wave(t);
 
-        t += frequency / sample_rate;
+        t += frequency / sample_rate as f32;
         t %= 1.;
 
         Some(value)
